Move common/colors.go into pkg/color
